requirements.txt
annotated-types==0.5.0
anyio==4.0.0
asgiref==3.7.2
assemblyai==0.18.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
decorator==4.4.2
Django==4.2.5
djangorestframework==3.14.0
drf-yasg==1.21.7
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
imageio==2.31.4
imageio-ffmpeg==0.4.9
inflection==0.5.1
moviepy==1.0.3
numpy==1.26.0
packaging==23.2
Pillow==10.0.1
proglog==0.1.10
pydantic==2.4.2
pydantic_core==2.10.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
websockets==11.0.3
